Move CalculatorViewController to ViewController
